+++ sway-lsp/src/capabilities/rename.rs
@@ -39,7 +39,7 @@ pub fn prepare_rename(
 ) -> Option<PrepareRenameResponse> {
     let (ident, token) = session.token_map().token_at_position(&url, position)?;
     match token.parsed {
-        AstToken::Reassignment(_) => None,
+        AstToken::ReassignmentExpression(_) => None,
         _ => Some(PrepareRenameResponse::RangeWithPlaceholder {
             range: get_range_from_span(&ident.span()),
             placeholder: ident.as_str().to_string(),
