+++ sway-lsp/src/traverse/lexed_tree.rs
@@ -1,9 +1,6 @@
 use crate::{
-    core::{
-        token::{to_ident_key, AstToken, SymbolKind, Token},
-        token_map::TokenMap,
-    },
-    traverse::Parse,
+    core::token::{to_ident_key, AstToken, SymbolKind, Token},
+    traverse::{Parse, ParseContext},
 };
 use std::ops::ControlFlow;
 use sway_ast::{
@@ -16,119 +13,109 @@ use sway_ast::{
 use sway_core::language::lexed::LexedProgram;
 use sway_types::{Ident, Span, Spanned};
 
-pub struct LexedTree<'a> {
-    tokens: &'a TokenMap,
-}
-
-impl<'a> LexedTree<'a> {
-    pub fn new(tokens: &'a TokenMap) -> Self {
-        Self { tokens }
+pub fn parse(lexed_program: &LexedProgram, ctx: &ParseContext) {
+    insert_keyword(ctx, lexed_program.root.tree.kind.span());
+    for item in &lexed_program.root.tree.items {
+        item.value.parse(ctx);
     }
 
-    pub fn parse(&self, lexed_program: &LexedProgram) {
-        insert_keyword(self.tokens, lexed_program.root.tree.kind.span());
-        for item in &lexed_program.root.tree.items {
-            item.value.parse(self.tokens);
-        }
-
-        for (.., dep) in &lexed_program.root.submodules {
-            insert_keyword(self.tokens, dep.module.tree.kind.span());
-            for item in &dep.module.tree.items {
-                item.value.parse(self.tokens);
-            }
+    for (.., dep) in &lexed_program.root.submodules {
+        insert_keyword(ctx, dep.module.tree.kind.span());
+        for item in &dep.module.tree.items {
+            item.value.parse(ctx);
         }
     }
 }
 
-fn insert_keyword(tokens: &TokenMap, span: Span) {
+fn insert_keyword(ctx: &ParseContext, span: Span) {
     let ident = Ident::new(span);
     let token = Token::from_parsed(AstToken::Keyword(ident.clone()), SymbolKind::Keyword);
-    tokens.insert(to_ident_key(&ident), token);
+    ctx.tokens.insert(to_ident_key(&ident), token);
 }
 
 impl Parse for ItemKind {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         match self {
             ItemKind::Dependency(dependency) => {
-                insert_keyword(tokens, dependency.dep_token.span());
+                insert_keyword(ctx, dependency.dep_token.span());
             }
             ItemKind::Use(item_use) => {
-                item_use.parse(tokens);
+                item_use.parse(ctx);
             }
             ItemKind::Struct(item_struct) => {
-                item_struct.parse(tokens);
+                item_struct.parse(ctx);
             }
             ItemKind::Enum(item_enum) => {
-                item_enum.parse(tokens);
+                item_enum.parse(ctx);
             }
             ItemKind::Fn(item_func) => {
-                item_func.parse(tokens);
+                item_func.parse(ctx);
             }
             ItemKind::Trait(item_trait) => {
-                item_trait.parse(tokens);
+                item_trait.parse(ctx);
             }
             ItemKind::Impl(item_impl) => {
-                item_impl.parse(tokens);
+                item_impl.parse(ctx);
             }
             ItemKind::Abi(item_abi) => {
-                item_abi.parse(tokens);
+                item_abi.parse(ctx);
             }
             ItemKind::Const(item_const) => {
-                item_const.parse(tokens);
+                item_const.parse(ctx);
             }
             ItemKind::Storage(item_storage) => {
-                item_storage.parse(tokens);
+                item_storage.parse(ctx);
             }
         }
     }
 }
 
 impl Parse for Expr {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         match self {
             Expr::AbiCast { abi_token, args } => {
-                insert_keyword(tokens, abi_token.span());
-                args.get().address.parse(tokens);
+                insert_keyword(ctx, abi_token.span());
+                args.get().address.parse(ctx);
             }
             Expr::Struct { fields, .. } => {
                 for expr in fields.get() {
-                    expr.parse(tokens);
+                    expr.parse(ctx);
                 }
             }
             Expr::Tuple(tuple) => {
-                tuple.get().parse(tokens);
+                tuple.get().parse(ctx);
             }
             Expr::Parens(parens) => {
-                parens.get().parse(tokens);
+                parens.get().parse(ctx);
             }
             Expr::Block(block) => {
-                block.get().parse(tokens);
+                block.get().parse(ctx);
             }
             Expr::Array(array) => {
-                array.get().parse(tokens);
+                array.get().parse(ctx);
             }
             Expr::Return {
                 return_token,
                 expr_opt,
             } => {
-                insert_keyword(tokens, return_token.span());
+                insert_keyword(ctx, return_token.span());
                 if let Some(expr) = expr_opt {
-                    expr.parse(tokens);
+                    expr.parse(ctx);
                 }
             }
             Expr::If(if_expr) => {
-                if_expr.parse(tokens);
+                if_expr.parse(ctx);
             }
             Expr::Match {
                 match_token,
                 value,
                 branches,
             } => {
-                insert_keyword(tokens, match_token.span());
-                value.parse(tokens);
+                insert_keyword(ctx, match_token.span());
+                value.parse(ctx);
                 for branch in branches.get() {
-                    branch.pattern.parse(tokens);
-                    branch.kind.parse(tokens);
+                    branch.pattern.parse(ctx);
+                    branch.kind.parse(ctx);
                 }
             }
             Expr::While {
@@ -136,19 +123,19 @@ impl Parse for Expr {
                 condition,
                 block,
             } => {
-                insert_keyword(tokens, while_token.span());
-                condition.parse(tokens);
-                block.get().parse(tokens);
+                insert_keyword(ctx, while_token.span());
+                condition.parse(ctx);
+                block.get().parse(ctx);
             }
             Expr::FuncApp { func, args } => {
-                func.parse(tokens);
+                func.parse(ctx);
                 for expr in args.get().into_iter() {
-                    expr.parse(tokens);
+                    expr.parse(ctx);
                 }
             }
             Expr::Index { target, arg } => {
-                target.parse(tokens);
-                arg.get().parse(tokens);
+                target.parse(ctx);
+                arg.get().parse(ctx);
             }
             Expr::MethodCall {
                 target,
@@ -156,32 +143,32 @@ impl Parse for Expr {
                 args,
                 ..
             } => {
-                target.parse(tokens);
+                target.parse(ctx);
                 if let Some(contract_args) = contract_args_opt {
                     for expr in contract_args.get().into_iter() {
-                        expr.parse(tokens);
+                        expr.parse(ctx);
                     }
                 }
                 for expr in args.get().into_iter() {
-                    expr.parse(tokens);
+                    expr.parse(ctx);
                 }
             }
             Expr::FieldProjection { target, .. } => {
-                target.parse(tokens);
+                target.parse(ctx);
             }
             Expr::TupleFieldProjection { target, .. } => {
-                target.parse(tokens);
+                target.parse(ctx);
             }
             Expr::Ref { ref_token, expr } => {
-                insert_keyword(tokens, ref_token.span());
-                expr.parse(tokens);
+                insert_keyword(ctx, ref_token.span());
+                expr.parse(ctx);
             }
             Expr::Deref { deref_token, expr } => {
-                insert_keyword(tokens, deref_token.span());
-                expr.parse(tokens);
+                insert_keyword(ctx, deref_token.span());
+                expr.parse(ctx);
             }
             Expr::Not { expr, .. } => {
-                expr.parse(tokens);
+                expr.parse(ctx);
             }
             Expr::Mul { lhs, rhs, .. }
             | Expr::Div { lhs, rhs, .. }
@@ -202,20 +189,20 @@ impl Parse for Expr {
             | Expr::GreaterThanEq { lhs, rhs, .. }
             | Expr::LogicalAnd { lhs, rhs, .. }
             | Expr::LogicalOr { lhs, rhs, .. } => {
-                lhs.parse(tokens);
-                rhs.parse(tokens);
+                lhs.parse(ctx);
+                rhs.parse(ctx);
             }
             Expr::Reassignment {
                 assignable, expr, ..
             } => {
-                assignable.parse(tokens);
-                expr.parse(tokens);
+                assignable.parse(ctx);
+                expr.parse(ctx);
             }
             Expr::Break { break_token } => {
-                insert_keyword(tokens, break_token.span());
+                insert_keyword(ctx, break_token.span());
             }
             Expr::Continue { continue_token } => {
-                insert_keyword(tokens, continue_token.span());
+                insert_keyword(ctx, continue_token.span());
             }
             _ => {}
         }
@@ -223,167 +210,167 @@ impl Parse for Expr {
 }
 
 impl Parse for ItemUse {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         if let Some(visibility) = &self.visibility {
-            insert_keyword(tokens, visibility.span());
+            insert_keyword(ctx, visibility.span());
         }
-        insert_keyword(tokens, self.use_token.span());
-        self.tree.parse(tokens);
+        insert_keyword(ctx, self.use_token.span());
+        self.tree.parse(ctx);
     }
 }
 
 impl Parse for ItemStruct {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         if let Some(visibility) = &self.visibility {
-            insert_keyword(tokens, visibility.span());
+            insert_keyword(ctx, visibility.span());
         }
-        insert_keyword(tokens, self.struct_token.span());
+        insert_keyword(ctx, self.struct_token.span());
 
         if let Some(where_clause_opt) = &self.where_clause_opt {
-            insert_keyword(tokens, where_clause_opt.where_token.span());
+            insert_keyword(ctx, where_clause_opt.where_token.span());
         }
 
         self.fields
             .get()
             .into_iter()
-            .for_each(|field| field.value.parse(tokens));
+            .for_each(|field| field.value.parse(ctx));
     }
 }
 
 impl Parse for ItemEnum {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         if let Some(visibility) = &self.visibility {
-            insert_keyword(tokens, visibility.span());
+            insert_keyword(ctx, visibility.span());
         }
-        insert_keyword(tokens, self.enum_token.span());
+        insert_keyword(ctx, self.enum_token.span());
 
         if let Some(where_clause_opt) = &self.where_clause_opt {
-            insert_keyword(tokens, where_clause_opt.where_token.span());
+            insert_keyword(ctx, where_clause_opt.where_token.span());
         }
 
         self.fields
             .get()
             .into_iter()
-            .for_each(|field| field.value.parse(tokens));
+            .for_each(|field| field.value.parse(ctx));
     }
 }
 
 impl Parse for ItemFn {
-    fn parse(&self, tokens: &TokenMap) {
-        self.fn_signature.parse(tokens);
-        self.body.get().parse(tokens);
+    fn parse(&self, ctx: &ParseContext) {
+        self.fn_signature.parse(ctx);
+        self.body.get().parse(ctx);
     }
 }
 
 impl Parse for ItemTrait {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         if let Some(visibility) = &self.visibility {
-            insert_keyword(tokens, visibility.span());
+            insert_keyword(ctx, visibility.span());
         }
-        insert_keyword(tokens, self.trait_token.span());
+        insert_keyword(ctx, self.trait_token.span());
 
         if let Some(where_clause_opt) = &self.where_clause_opt {
-            insert_keyword(tokens, where_clause_opt.where_token.span());
+            insert_keyword(ctx, where_clause_opt.where_token.span());
         }
 
         self.trait_items
             .get()
             .iter()
-            .for_each(|item| item.0.value.parse(tokens));
+            .for_each(|item| item.0.value.parse(ctx));
 
         if let Some(trait_defs_opt) = &self.trait_defs_opt {
             trait_defs_opt
                 .get()
                 .iter()
-                .for_each(|item| item.value.parse(tokens));
+                .for_each(|item| item.value.parse(ctx));
         }
     }
 }
 
 impl Parse for ItemImpl {
-    fn parse(&self, tokens: &TokenMap) {
-        insert_keyword(tokens, self.impl_token.span());
+    fn parse(&self, ctx: &ParseContext) {
+        insert_keyword(ctx, self.impl_token.span());
 
         if let Some((.., for_token)) = &self.trait_opt {
-            insert_keyword(tokens, for_token.span());
+            insert_keyword(ctx, for_token.span());
         }
 
-        self.ty.parse(tokens);
+        self.ty.parse(ctx);
 
         if let Some(where_clause_opt) = &self.where_clause_opt {
-            insert_keyword(tokens, where_clause_opt.where_token.span());
+            insert_keyword(ctx, where_clause_opt.where_token.span());
         }
 
         self.contents
             .get()
             .iter()
-            .for_each(|item| item.value.parse(tokens));
+            .for_each(|item| item.value.parse(ctx));
     }
 }
 
 impl Parse for ItemAbi {
-    fn parse(&self, tokens: &TokenMap) {
-        insert_keyword(tokens, self.abi_token.span());
+    fn parse(&self, ctx: &ParseContext) {
+        insert_keyword(ctx, self.abi_token.span());
 
         self.abi_items.get().iter().for_each(|item| {
-            item.0.value.parse(tokens);
+            item.0.value.parse(ctx);
         });
 
         if let Some(abi_defs_opt) = self.abi_defs_opt.as_ref() {
             abi_defs_opt
                 .get()
                 .iter()
-                .for_each(|item| item.value.parse(tokens));
+                .for_each(|item| item.value.parse(ctx));
         }
     }
 }
 
 impl Parse for ItemConst {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         if let Some(visibility) = &self.visibility {
-            insert_keyword(tokens, visibility.span());
+            insert_keyword(ctx, visibility.span());
         }
-        insert_keyword(tokens, self.const_token.span());
+        insert_keyword(ctx, self.const_token.span());
 
         if let Some((.., ty)) = self.ty_opt.as_ref() {
-            ty.parse(tokens);
+            ty.parse(ctx);
         }
 
-        self.expr.parse(tokens);
+        self.expr.parse(ctx);
     }
 }
 
 impl Parse for ItemStorage {
-    fn parse(&self, tokens: &TokenMap) {
-        insert_keyword(tokens, self.storage_token.span());
+    fn parse(&self, ctx: &ParseContext) {
+        insert_keyword(ctx, self.storage_token.span());
 
         self.fields
             .get()
             .into_iter()
-            .for_each(|field| field.value.parse(tokens));
+            .for_each(|field| field.value.parse(ctx));
     }
 }
 
 impl Parse for StorageField {
-    fn parse(&self, tokens: &TokenMap) {
-        self.ty.parse(tokens);
-        self.initializer.parse(tokens);
+    fn parse(&self, ctx: &ParseContext) {
+        self.ty.parse(ctx);
+        self.initializer.parse(ctx);
     }
 }
 
 impl Parse for UseTree {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         match self {
             UseTree::Group { imports } => {
                 for use_tree in imports.get().into_iter() {
-                    use_tree.parse(tokens);
+                    use_tree.parse(ctx);
                 }
             }
             UseTree::Rename { as_token, .. } => {
-                insert_keyword(tokens, as_token.span());
+                insert_keyword(ctx, as_token.span());
             }
             UseTree::Path { suffix, .. } => {
-                suffix.parse(tokens);
+                suffix.parse(ctx);
             }
             _ => {}
         }
@@ -391,25 +378,25 @@ impl Parse for UseTree {
 }
 
 impl Parse for TypeField {
-    fn parse(&self, tokens: &TokenMap) {
-        self.ty.parse(tokens);
+    fn parse(&self, ctx: &ParseContext) {
+        self.ty.parse(ctx);
     }
 }
 
 impl Parse for Ty {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         match self {
             Ty::Tuple(tuple) => {
-                tuple.get().parse(tokens);
+                tuple.get().parse(ctx);
             }
             Ty::Array(array) => {
                 let inner = array.get();
-                inner.ty.parse(tokens);
-                inner.length.parse(tokens);
+                inner.ty.parse(ctx);
+                inner.length.parse(ctx);
             }
             Ty::Str { str_token, length } => {
-                insert_keyword(tokens, str_token.span());
-                length.get().parse(tokens);
+                insert_keyword(ctx, str_token.span());
+                length.get().parse(ctx);
             }
             _ => {}
         }
@@ -417,27 +404,27 @@ impl Parse for Ty {
 }
 
 impl Parse for FnSignature {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         if let Some(visibility) = &self.visibility {
-            insert_keyword(tokens, visibility.span());
+            insert_keyword(ctx, visibility.span());
         }
-        insert_keyword(tokens, self.fn_token.span());
+        insert_keyword(ctx, self.fn_token.span());
 
-        self.arguments.get().parse(tokens);
+        self.arguments.get().parse(ctx);
         if let Some((.., ty)) = &self.return_type_opt {
-            ty.parse(tokens);
+            ty.parse(ctx);
         }
         if let Some(where_clause) = &self.where_clause_opt {
-            insert_keyword(tokens, where_clause.where_token.span());
+            insert_keyword(ctx, where_clause.where_token.span());
         }
     }
 }
 
 impl Parse for FnArgs {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         match self {
             FnArgs::Static(punct) => {
-                punct.into_iter().for_each(|fn_arg| fn_arg.parse(tokens));
+                punct.into_iter().for_each(|fn_arg| fn_arg.parse(ctx));
             }
             FnArgs::NonStatic {
                 self_token,
@@ -445,15 +432,15 @@ impl Parse for FnArgs {
                 mutable_self,
                 args_opt,
             } => {
-                insert_keyword(tokens, self_token.span());
+                insert_keyword(ctx, self_token.span());
                 if let Some(ref_token) = ref_self {
-                    insert_keyword(tokens, ref_token.span());
+                    insert_keyword(ctx, ref_token.span());
                 }
                 if let Some(mut_token) = mutable_self {
-                    insert_keyword(tokens, mut_token.span());
+                    insert_keyword(ctx, mut_token.span());
                 }
                 if let Some((.., punct)) = args_opt {
-                    punct.into_iter().for_each(|fn_arg| fn_arg.parse(tokens));
+                    punct.into_iter().for_each(|fn_arg| fn_arg.parse(ctx));
                 }
             }
         }
@@ -461,79 +448,79 @@ impl Parse for FnArgs {
 }
 
 impl Parse for FnArg {
-    fn parse(&self, tokens: &TokenMap) {
-        self.pattern.parse(tokens);
-        self.ty.parse(tokens);
+    fn parse(&self, ctx: &ParseContext) {
+        self.pattern.parse(ctx);
+        self.ty.parse(ctx);
     }
 }
 
 impl Parse for CodeBlockContents {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         for statement in self.statements.iter() {
-            statement.parse(tokens);
+            statement.parse(ctx);
         }
         if let Some(expr) = self.final_expr_opt.as_ref() {
-            expr.parse(tokens);
+            expr.parse(ctx);
         }
     }
 }
 
 impl Parse for Statement {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         match self {
             Statement::Let(let_stmt) => {
-                let_stmt.parse(tokens);
+                let_stmt.parse(ctx);
             }
             Statement::Expr { expr, .. } => {
-                expr.parse(tokens);
+                expr.parse(ctx);
             }
             Statement::Item(item) => {
-                item.value.parse(tokens);
+                item.value.parse(ctx);
             }
         }
     }
 }
 
 impl Parse for StatementLet {
-    fn parse(&self, tokens: &TokenMap) {
-        insert_keyword(tokens, self.let_token.span());
-        self.pattern.parse(tokens);
+    fn parse(&self, ctx: &ParseContext) {
+        insert_keyword(ctx, self.let_token.span());
+        self.pattern.parse(ctx);
         if let Some((.., ty)) = &self.ty_opt {
-            ty.parse(tokens);
+            ty.parse(ctx);
         }
-        self.expr.parse(tokens);
+        self.expr.parse(ctx);
     }
 }
 
 impl Parse for ExprArrayDescriptor {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         match self {
             ExprArrayDescriptor::Sequence(punct) => {
                 for expr in punct.into_iter() {
-                    expr.parse(tokens);
+                    expr.parse(ctx);
                 }
             }
             ExprArrayDescriptor::Repeat { value, length, .. } => {
-                value.parse(tokens);
-                length.parse(tokens);
+                value.parse(ctx);
+                length.parse(ctx);
             }
         }
     }
 }
 
 impl Parse for IfExpr {
-    fn parse(&self, tokens: &TokenMap) {
-        insert_keyword(tokens, self.if_token.span());
-        self.condition.parse(tokens);
-        self.then_block.get().parse(tokens);
+    fn parse(&self, ctx: &ParseContext) {
+        insert_keyword(ctx, self.if_token.span());
+        self.condition.parse(ctx);
+        self.then_block.get().parse(ctx);
         if let Some((else_token, control_flow)) = &self.else_opt {
-            insert_keyword(tokens, else_token.span());
+            insert_keyword(ctx, else_token.span());
             match control_flow {
                 ControlFlow::Break(block) => {
-                    block.get().parse(tokens);
+                    block.get().parse(ctx);
                 }
                 ControlFlow::Continue(if_expr) => {
-                    if_expr.parse(tokens);
+                    if_expr.parse(ctx);
                 }
             }
         }
@@ -541,10 +528,10 @@ impl Parse for IfExpr {
 }
 
 impl Parse for IfCondition {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         match self {
             IfCondition::Expr(expr) => {
-                expr.parse(tokens);
+                expr.parse(ctx);
             }
             IfCondition::Let {
                 let_token,
@@ -552,35 +539,35 @@ impl Parse for IfCondition {
                 rhs,
                 ..
             } => {
-                insert_keyword(tokens, let_token.span());
-                lhs.parse(tokens);
-                rhs.parse(tokens);
+                insert_keyword(ctx, let_token.span());
+                lhs.parse(ctx);
+                rhs.parse(ctx);
             }
         }
     }
 }
 
 impl Parse for Pattern {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         match self {
             Pattern::Var {
                 reference, mutable, ..
             } => {
                 if let Some(reference) = reference {
-                    insert_keyword(tokens, reference.span());
+                    insert_keyword(ctx, reference.span());
                 }
                 if let Some(mutable) = mutable {
-                    insert_keyword(tokens, mutable.span());
+                    insert_keyword(ctx, mutable.span());
                 }
             }
             Pattern::Constructor { args, .. } | Pattern::Tuple(args) => {
                 for pattern in args.get().into_iter() {
-                    pattern.parse(tokens);
+                    pattern.parse(ctx);
                 }
             }
             Pattern::Struct { fields, .. } => {
                 for field in fields.get().into_iter() {
-                    field.parse(tokens);
+                    field.parse(ctx);
                 }
             }
             _ => {}
@@ -589,70 +576,70 @@ impl Parse for Pattern {
 }
 
 impl Parse for PatternStructField {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         if let PatternStructField::Field {
             pattern_opt: Some((.., pattern)),
             ..
         } = self
         {
-            pattern.parse(tokens);
+            pattern.parse(ctx);
         }
     }
 }
 
 impl Parse for MatchBranchKind {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         match self {
             MatchBranchKind::Block { block, .. } => {
-                block.get().parse(tokens);
+                block.get().parse(ctx);
             }
             MatchBranchKind::Expr { expr, .. } => {
-                expr.parse(tokens);
+                expr.parse(ctx);
             }
         }
     }
 }
 
 impl Parse for ExprStructField {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         if let Some((.., expr)) = &self.expr_opt {
-            expr.parse(tokens);
+            expr.parse(ctx);
         }
     }
 }
 
 impl Parse for ExprTupleDescriptor {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         if let ExprTupleDescriptor::Cons { head, tail, .. } = self {
-            head.parse(tokens);
+            head.parse(ctx);
             for expr in tail.into_iter() {
-                expr.parse(tokens);
+                expr.parse(ctx);
             }
         }
     }
 }
 
 impl Parse for TyTupleDescriptor {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         if let TyTupleDescriptor::Cons { head, tail, .. } = self {
-            head.parse(tokens);
+            head.parse(ctx);
             for expr in tail.into_iter() {
-                expr.parse(tokens);
+                expr.parse(ctx);
             }
         }
     }
 }
 
 impl Parse for Assignable {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         match self {
             Assignable::Index { target, arg } => {
-                target.parse(tokens);
-                arg.get().parse(tokens)
+                target.parse(ctx);
+                arg.get().parse(ctx)
             }
             Assignable::FieldProjection { target, .. }
             | Assignable::TupleFieldProjection { target, .. } => {
-                target.parse(tokens);
+                target.parse(ctx);
             }
             _ => {}
         }
