+++ sway-lsp/src/traverse/parsed_tree.rs
@@ -1,941 +1,771 @@
-#![allow(dead_code)]
-use std::iter;
-
 use crate::{
-    core::{
-        token::{
-            desugared_op, to_ident_key, type_info_to_symbol_kind, AstToken, SymbolKind, Token,
-            TypeDefinition,
-        },
-        token_map::TokenMap,
-    },
-    traverse::Parse,
+    core::token::{desugared_op, AstToken, SymbolKind},
+    traverse::{Parse, ParseContext},
 };
-
 use sway_core::{
     language::{
         parsed::{
-            AbiCastExpression, AmbiguousPathExpression, ArrayIndexExpression, AstNode,
-            AstNodeContent, CodeBlock, Declaration, DelineatedPathExpression, Expression,
+            AbiCastExpression, AbiDeclaration, AmbiguousPathExpression, ArrayExpression,
+            ArrayIndexExpression, AstNode, AstNodeContent, CodeBlock, ConstantDeclaration,
+            Declaration, DelineatedPathExpression, EnumDeclaration, EnumVariant, Expression,
             ExpressionKind, FunctionApplicationExpression, FunctionDeclaration, FunctionParameter,
-            IfExpression, IntrinsicFunctionExpression, LazyOperatorExpression, MatchExpression,
-            MethodApplicationExpression, MethodName, ReassignmentTarget, Scrutinee,
-            StorageAccessExpression, StructExpression, StructScrutineeField, SubfieldExpression,
-            TraitFn, TupleIndexExpression, WhileLoopExpression,
+            ImplSelf, ImplTrait, IntrinsicFunctionExpression, LazyOperatorExpression, MatchBranch,
+            MethodApplicationExpression, MethodName, ReassignmentExpression, ReassignmentTarget,
+            Scrutinee, StorageDeclaration, StorageField, StructDeclaration, StructExpression,
+            StructExpressionField, StructField, StructScrutineeField, SubfieldExpression,
+            Supertrait, TraitDeclaration, TraitFn, VariableDeclaration,
         },
         Literal,
     },
     transform::{AttributeKind, AttributesMap},
     type_system::{TypeArgument, TypeParameter},
-    TypeEngine, TypeInfo,
+    TypeInfo,
 };
 use sway_types::constants::{DESTRUCTURE_PREFIX, MATCH_RETURN_VAR_NAME_PREFIX, TUPLE_NAME_PREFIX};
-use sway_types::{Ident, Span, Spanned};
+use sway_types::{Ident, Spanned};
 
-pub struct ParsedTree<'a> {
-    type_engine: &'a TypeEngine,
-    tokens: &'a TokenMap,
+pub fn parse(node: &AstNode, ctx: &ParseContext) {
+    node.parse(ctx);
 }
 
-impl<'a> ParsedTree<'a> {
-    pub fn new(type_engine: &'a TypeEngine, tokens: &'a TokenMap) -> Self {
-        Self {
-            type_engine,
-            tokens,
-        }
-    }
-
-    pub fn traverse_node(&self, node: &AstNode) {
-        match &node.content {
-            AstNodeContent::Declaration(declaration) => self.handle_declaration(declaration),
+impl Parse for AstNode {
+    fn parse(&self, ctx: &ParseContext) {
+        match &self.content {
+            AstNodeContent::Declaration(declaration) => declaration.parse(ctx),
             AstNodeContent::Expression(expression)
-            | AstNodeContent::ImplicitReturnExpression(expression) => {
-                self.handle_expression(expression)
-            }
+            | AstNodeContent::ImplicitReturnExpression(expression) => expression.parse(ctx),
             // TODO
             // handle other content types
             _ => {}
         };
     }
+}
 
-    fn handle_function_declation(&self, func: &FunctionDeclaration) {
-        let token = Token::from_parsed(
-            AstToken::FunctionDeclaration(func.clone()),
-            SymbolKind::Function,
-        );
-        self.tokens.insert(to_ident_key(&func.name), token.clone());
-        for node in &func.body.contents {
-            self.traverse_node(node);
-        }
-
-        for parameter in &func.parameters {
-            self.collect_function_parameter(parameter);
-        }
-
-        for type_param in &func.type_parameters {
-            self.collect_type_parameter(type_param, AstToken::FunctionDeclaration(func.clone()));
-        }
-
-        self.collect_type_info_token(
-            &token,
-            &func.return_type,
-            Some(func.return_type_span.clone()),
-            None,
-        );
-
-        func.attributes.parse(self.tokens);
-    }
-
-    fn handle_declaration(&self, declaration: &Declaration) {
-        match declaration {
-            Declaration::VariableDeclaration(variable) => {
-                // Don't collect tokens if the ident's name contains __tuple_ || __match_return_var_name_
-                // The individual elements are handled in the subsequent VariableDeclaration's
-                if !variable.name.as_str().contains(TUPLE_NAME_PREFIX)
-                    && !variable
-                        .name
-                        .as_str()
-                        .contains(MATCH_RETURN_VAR_NAME_PREFIX)
-                {
-                    let symbol_kind = if variable.name.as_str().contains(DESTRUCTURE_PREFIX) {
-                        SymbolKind::Struct
-                    } else {
-                        SymbolKind::Variable
-                    };
-
-                    let token =
-                        Token::from_parsed(AstToken::Declaration(declaration.clone()), symbol_kind);
-                    // We want to use the span from variable.name to construct a
-                    // new Ident as the name_override_opt can be set to one of the
-                    // const prefixes and not the actual token name.
-                    self.tokens.insert(
-                        to_ident_key(&Ident::new(variable.name.span())),
-                        token.clone(),
-                    );
-
-                    if let Some(type_ascription_span) = &variable.type_ascription_span {
-                        self.collect_type_info_token(
-                            &token,
-                            &variable.type_ascription,
-                            Some(type_ascription_span.clone()),
-                            None,
-                        );
-                    }
-                }
-                self.handle_expression(&variable.body);
-            }
-            Declaration::FunctionDeclaration(func) => {
-                self.handle_function_declation(func);
-            }
-            Declaration::TraitDeclaration(trait_decl) => {
-                self.tokens.insert(
-                    to_ident_key(&trait_decl.name),
-                    Token::from_parsed(
-                        AstToken::Declaration(declaration.clone()),
-                        SymbolKind::Trait,
-                    ),
-                );
-
-                for trait_fn in &trait_decl.interface_surface {
-                    self.collect_trait_fn(trait_fn);
-                }
-
-                for func_dec in &trait_decl.methods {
-                    self.handle_function_declation(func_dec);
-                }
-
-                for supertrait in &trait_decl.supertraits {
-                    self.tokens.insert(
-                        to_ident_key(&supertrait.name.suffix),
-                        Token::from_parsed(
-                            AstToken::Declaration(declaration.clone()),
-                            SymbolKind::Trait,
-                        ),
-                    );
-                }
-            }
-            Declaration::StructDeclaration(struct_dec) => {
-                self.tokens.insert(
-                    to_ident_key(&struct_dec.name),
-                    Token::from_parsed(
-                        AstToken::Declaration(declaration.clone()),
-                        SymbolKind::Struct,
-                    ),
-                );
-                for field in &struct_dec.fields {
-                    let token =
-                        Token::from_parsed(AstToken::StructField(field.clone()), SymbolKind::Field);
-                    self.tokens.insert(to_ident_key(&field.name), token.clone());
-
-                    self.collect_type_info_token(
-                        &token,
-                        &field.type_info,
-                        Some(field.type_span.clone()),
-                        None,
-                    );
-
-                    field.attributes.parse(self.tokens);
-                }
-
-                for type_param in &struct_dec.type_parameters {
-                    self.collect_type_parameter(
-                        type_param,
-                        AstToken::Declaration(declaration.clone()),
-                    );
-                }
-
-                struct_dec.attributes.parse(self.tokens);
-            }
-            Declaration::EnumDeclaration(enum_decl) => {
-                self.tokens.insert(
-                    to_ident_key(&enum_decl.name),
-                    Token::from_parsed(
-                        AstToken::Declaration(declaration.clone()),
-                        SymbolKind::Enum,
-                    ),
-                );
-
-                for type_param in &enum_decl.type_parameters {
-                    self.collect_type_parameter(
-                        type_param,
-                        AstToken::Declaration(declaration.clone()),
-                    );
-                }
-
-                for variant in &enum_decl.variants {
-                    let token = Token::from_parsed(
-                        AstToken::EnumVariant(variant.clone()),
-                        SymbolKind::Variant,
-                    );
-                    self.tokens
-                        .insert(to_ident_key(&variant.name), token.clone());
-
-                    self.collect_type_info_token(
-                        &token,
-                        &variant.type_info,
-                        Some(variant.type_span.clone()),
-                        Some(SymbolKind::Variant),
-                    );
-                    variant.attributes.parse(self.tokens);
-                }
-
-                enum_decl.attributes.parse(self.tokens);
-            }
-            Declaration::ImplTrait(impl_trait) => {
-                for ident in &impl_trait.trait_name.prefixes {
-                    self.tokens.insert(
-                        to_ident_key(ident),
-                        Token::from_parsed(
-                            AstToken::Declaration(declaration.clone()),
-                            SymbolKind::Module,
-                        ),
-                    );
-                }
-
-                self.tokens.insert(
-                    to_ident_key(&impl_trait.trait_name.suffix),
-                    Token::from_parsed(
-                        AstToken::Declaration(declaration.clone()),
-                        SymbolKind::Trait,
-                    ),
-                );
-
-                let token = Token::from_parsed(
-                    AstToken::Declaration(declaration.clone()),
-                    type_info_to_symbol_kind(self.type_engine, &impl_trait.type_implementing_for),
-                );
-
-                self.collect_type_info_token(
-                    &token,
-                    &impl_trait.type_implementing_for,
-                    Some(impl_trait.type_implementing_for_span.clone()),
-                    Some(SymbolKind::Variant),
-                );
-
-                for type_param in &impl_trait.impl_type_parameters {
-                    self.collect_type_parameter(
-                        type_param,
-                        AstToken::Declaration(declaration.clone()),
-                    );
-                }
-
-                for func_dec in &impl_trait.functions {
-                    self.handle_function_declation(func_dec);
-                }
-            }
-            Declaration::ImplSelf(impl_self) => {
-                if let TypeInfo::Custom {
-                    name,
-                    type_arguments,
-                } = &impl_self.type_implementing_for
-                {
-                    let token = Token::from_parsed(
-                        AstToken::Declaration(declaration.clone()),
-                        SymbolKind::Struct,
-                    );
-                    self.tokens.insert(to_ident_key(name), token.clone());
-                    if let Some(type_arguments) = type_arguments {
-                        for type_arg in type_arguments {
-                            self.collect_type_arg(type_arg, &token);
-                        }
-                    }
-                }
-
-                for type_param in &impl_self.impl_type_parameters {
-                    self.collect_type_parameter(
-                        type_param,
-                        AstToken::Declaration(declaration.clone()),
-                    );
-                }
-
-                for func_dec in &impl_self.functions {
-                    self.handle_function_declation(func_dec);
-                }
-            }
-            Declaration::AbiDeclaration(abi_decl) => {
-                self.tokens.insert(
-                    to_ident_key(&abi_decl.name),
-                    Token::from_parsed(
-                        AstToken::Declaration(declaration.clone()),
-                        SymbolKind::Trait,
-                    ),
-                );
-
-                for trait_fn in &abi_decl.interface_surface {
-                    self.collect_trait_fn(trait_fn);
-                }
-
-                abi_decl.attributes.parse(self.tokens);
-            }
-            Declaration::ConstantDeclaration(const_decl) => {
-                let token = Token::from_parsed(
-                    AstToken::Declaration(declaration.clone()),
-                    SymbolKind::Const,
-                );
-                self.tokens
-                    .insert(to_ident_key(&const_decl.name), token.clone());
-
-                self.collect_type_info_token(
-                    &token,
-                    &const_decl.type_ascription,
-                    const_decl.type_ascription_span.clone(),
-                    None,
-                );
-                self.handle_expression(&const_decl.value);
-
-                const_decl.attributes.parse(self.tokens);
-            }
-            Declaration::StorageDeclaration(storage_decl) => {
-                for field in &storage_decl.fields {
-                    let token = Token::from_parsed(
-                        AstToken::StorageField(field.clone()),
-                        SymbolKind::Field,
-                    );
-                    self.tokens.insert(to_ident_key(&field.name), token.clone());
-
-                    self.collect_type_info_token(
-                        &token,
-                        &field.type_info,
-                        Some(field.type_info_span.clone()),
-                        None,
-                    );
-                    self.handle_expression(&field.initializer);
-
-                    field.attributes.parse(self.tokens);
-                }
-                storage_decl.attributes.parse(self.tokens);
-            }
+impl Parse for Declaration {
+    fn parse(&self, ctx: &ParseContext) {
+        match self {
+            Declaration::VariableDeclaration(variable) => variable.parse(ctx),
+            Declaration::FunctionDeclaration(func_decl) => func_decl.parse(ctx),
+            Declaration::TraitDeclaration(trait_decl) => trait_decl.parse(ctx),
+            Declaration::StructDeclaration(struct_dec) => struct_dec.parse(ctx),
+            Declaration::EnumDeclaration(enum_decl) => enum_decl.parse(ctx),
+            Declaration::ImplTrait(impl_trait) => impl_trait.parse(ctx),
+            Declaration::ImplSelf(impl_self) => impl_self.parse(ctx),
+            Declaration::AbiDeclaration(abi_decl) => abi_decl.parse(ctx),
+            Declaration::ConstantDeclaration(const_decl) => const_decl.parse(ctx),
+            Declaration::StorageDeclaration(storage_decl) => storage_decl.parse(ctx),
         }
     }
+}
 
-    fn handle_expression(&self, expression: &Expression) {
-        let span = &expression.span;
-        match &expression.kind {
+impl Parse for Expression {
+    fn parse(&self, ctx: &ParseContext) {
+        match &self.kind {
             ExpressionKind::Error(_part_spans) => {
                 // FIXME(Centril): Left for @JoshuaBatty to use.
             }
             ExpressionKind::Literal(value) => {
-                let symbol_kind = literal_to_symbol_kind(value);
-
-                self.tokens.insert(
-                    to_ident_key(&Ident::new(span.clone())),
-                    Token::from_parsed(AstToken::Expression(expression.clone()), symbol_kind),
+                let symbol_kind = literal_to_symbol_kind(&value);
+                ctx.tokens.insert_parsed(
+                    Ident::new(self.span()),
+                    AstToken::Literal(value.clone()),
+                    symbol_kind,
                 );
             }
             ExpressionKind::FunctionApplication(function_application_expression) => {
-                let FunctionApplicationExpression {
-                    call_path_binding,
-                    arguments,
-                } = &**function_application_expression;
-                // Don't collect applications of desugared operators due to mismatched ident lengths.
-                if !desugared_op(&call_path_binding.inner.prefixes) {
-                    for ident in &call_path_binding.inner.prefixes {
-                        self.tokens.insert(
-                            to_ident_key(ident),
-                            Token::from_parsed(
-                                AstToken::Expression(expression.clone()),
-                                SymbolKind::Module,
-                            ),
-                        );
-                    }
-
-                    let token = Token::from_parsed(
-                        AstToken::Expression(expression.clone()),
-                        SymbolKind::Function,
-                    );
-
-                    self.tokens
-                        .insert(to_ident_key(&call_path_binding.inner.suffix), token.clone());
-
-                    for type_arg in &call_path_binding.type_arguments {
-                        self.collect_type_arg(type_arg, &token);
-                    }
-                }
-
-                for exp in arguments {
-                    self.handle_expression(exp);
-                }
+                function_application_expression.parse(ctx)
             }
             ExpressionKind::LazyOperator(LazyOperatorExpression { lhs, rhs, .. }) => {
-                self.handle_expression(lhs);
-                self.handle_expression(rhs);
+                lhs.parse(ctx);
+                rhs.parse(ctx);
             }
             ExpressionKind::Variable(name) => {
                 if !name.as_str().contains(TUPLE_NAME_PREFIX)
                     && !name.as_str().contains(MATCH_RETURN_VAR_NAME_PREFIX)
                 {
-                    let symbol_kind = if name.as_str().contains(DESTRUCTURE_PREFIX) {
-                        SymbolKind::Struct
-                    } else {
-                        SymbolKind::Variable
-                    };
-
-                    self.tokens.insert(
-                        to_ident_key(name),
-                        Token::from_parsed(AstToken::Expression(expression.clone()), symbol_kind),
+                    let symbol_kind = symbol_kind_from_ident(&name);
+                    ctx.tokens.insert_parsed(
+                        name.clone(),
+                        AstToken::Ident(name.clone()),
+                        symbol_kind,
                     );
                 }
             }
             ExpressionKind::Tuple(fields) => {
-                for exp in fields {
-                    self.handle_expression(exp);
-                }
+                fields.iter().for_each(|field| field.parse(ctx));
             }
-            ExpressionKind::TupleIndex(TupleIndexExpression {
-                prefix, index_span, ..
-            }) => {
-                self.handle_expression(prefix);
-
-                self.tokens.insert(
-                    to_ident_key(&Ident::new(index_span.clone())),
-                    Token::from_parsed(
-                        AstToken::Expression(expression.clone()),
-                        SymbolKind::NumericLiteral,
-                    ),
+            ExpressionKind::TupleIndex(tuple_index_exp) => {
+                tuple_index_exp.prefix.parse(ctx);
+                let ident = Ident::new(tuple_index_exp.index_span.clone());
+                ctx.tokens.insert_parsed(
+                    ident.clone(),
+                    AstToken::Ident(ident.clone()),
+                    SymbolKind::NumericLiteral,
                 );
             }
             ExpressionKind::Array(array_expression) => {
-                for exp in &array_expression.contents {
-                    self.handle_expression(exp);
-                }
-
-                if let Some(length_span) = &array_expression.length_span {
-                    self.tokens.insert(
-                        to_ident_key(&Ident::new(length_span.clone())),
-                        Token::from_parsed(
-                            AstToken::Expression(expression.clone()),
-                            SymbolKind::NumericLiteral,
-                        ),
-                    );
-                }
+                array_expression.parse(ctx);
             }
             ExpressionKind::Struct(struct_expression) => {
-                let StructExpression {
-                    call_path_binding,
-                    fields,
-                } = &**struct_expression;
-                for ident in &call_path_binding.inner.prefixes {
-                    self.tokens.insert(
-                        to_ident_key(ident),
-                        Token::from_parsed(
-                            AstToken::Expression(expression.clone()),
-                            SymbolKind::Struct,
-                        ),
-                    );
-                }
-
-                let name = &call_path_binding.inner.suffix;
-                let type_arguments = &call_path_binding.type_arguments;
-
-                let token = Token::from_parsed(
-                    AstToken::Expression(expression.clone()),
-                    SymbolKind::Struct,
-                );
-                self.tokens.insert(to_ident_key(name), token.clone());
-                for type_arg in type_arguments {
-                    self.collect_type_arg(type_arg, &token);
-                }
-
-                for field in fields {
-                    self.tokens.insert(
-                        to_ident_key(&field.name),
-                        Token::from_parsed(
-                            AstToken::StructExpressionField(field.clone()),
-                            SymbolKind::Field,
-                        ),
-                    );
-                    self.handle_expression(&field.value);
+                struct_expression.parse(ctx);
+            }
+            ExpressionKind::CodeBlock(code_block) => {
+                code_block.parse(ctx);
+            }
+            ExpressionKind::If(if_exp) => {
+                if_exp.condition.parse(ctx);
+                if_exp.then.parse(ctx);
+                if let Some(r#else) = &if_exp.r#else {
+                    r#else.parse(ctx);
                 }
             }
-            ExpressionKind::CodeBlock(contents) => {
-                for node in &contents.contents {
-                    self.traverse_node(node);
-                }
-            }
-            ExpressionKind::If(IfExpression {
-                condition,
-                then,
-                r#else,
-                ..
-            }) => {
-                self.handle_expression(condition);
-                self.handle_expression(then);
-                if let Some(r#else) = r#else {
-                    self.handle_expression(r#else);
-                }
-            }
-            ExpressionKind::Match(MatchExpression {
-                value, branches, ..
-            }) => {
-                self.handle_expression(value);
-                for branch in branches {
-                    self.collect_scrutinee(&branch.scrutinee);
-                    self.handle_expression(&branch.result);
-                }
+            ExpressionKind::Match(match_exp) => {
+                match_exp.value.parse(ctx);
+                match_exp
+                    .branches
+                    .iter()
+                    .for_each(|branch| branch.parse(ctx));
             }
             ExpressionKind::Asm(_) => {
                 //TODO handle asm expressions
             }
             ExpressionKind::MethodApplication(method_application_expression) => {
-                let MethodApplicationExpression {
-                    method_name_binding,
-                    arguments,
-                    contract_call_params,
-                } = &**method_application_expression;
-                let prefixes = match &method_name_binding.inner {
-                    MethodName::FromType {
-                        call_path_binding, ..
-                    } => call_path_binding.inner.prefixes.clone(),
-                    MethodName::FromTrait { call_path, .. } => call_path.prefixes.clone(),
-                    _ => vec![],
-                };
-
-                if let MethodName::FromType {
-                    call_path_binding, ..
-                } = &method_name_binding.inner
-                {
-                    let token = Token::from_parsed(
-                        AstToken::Expression(expression.clone()),
-                        SymbolKind::Struct,
-                    );
-                    let (type_info, span) = &call_path_binding.inner.suffix;
-                    self.collect_type_info_token(&token, type_info, Some(span.clone()), None);
-                }
-
-                // Don't collect applications of desugared operators due to mismatched ident lengths.
-                if !desugared_op(&prefixes) {
-                    self.tokens.insert(
-                        to_ident_key(&method_name_binding.inner.easy_name()),
-                        Token::from_parsed(
-                            AstToken::Expression(expression.clone()),
-                            SymbolKind::Struct,
-                        ),
-                    );
-                }
-
-                for exp in arguments {
-                    self.handle_expression(exp);
-                }
-
-                for field in contract_call_params {
-                    self.tokens.insert(
-                        to_ident_key(&field.name),
-                        Token::from_parsed(
-                            AstToken::Expression(field.value.clone()),
-                            SymbolKind::Field,
-                        ),
-                    );
-                    self.handle_expression(&field.value);
-                }
+                method_application_expression.parse(ctx);
             }
             ExpressionKind::Subfield(SubfieldExpression {
                 prefix,
                 field_to_access,
                 ..
             }) => {
-                self.tokens.insert(
-                    to_ident_key(field_to_access),
-                    Token::from_parsed(AstToken::Expression(expression.clone()), SymbolKind::Field),
+                ctx.tokens.insert_parsed(
+                    field_to_access.clone(),
+                    AstToken::Ident(field_to_access.clone()),
+                    SymbolKind::Field,
                 );
-                self.handle_expression(prefix);
+                prefix.parse(ctx);
             }
             ExpressionKind::AmbiguousPathExpression(path_expr) => {
-                let AmbiguousPathExpression {
-                    call_path_binding,
-                    args,
-                } = &**path_expr;
-
-                for ident in call_path_binding
-                    .inner
-                    .prefixes
-                    .iter()
-                    .chain(iter::once(&call_path_binding.inner.suffix.before.inner))
-                {
-                    self.tokens.insert(
-                        to_ident_key(ident),
-                        Token::from_parsed(
-                            AstToken::Expression(expression.clone()),
-                            SymbolKind::Enum,
-                        ),
-                    );
-                }
-
-                let token = Token::from_parsed(
-                    AstToken::Expression(expression.clone()),
-                    SymbolKind::Variant,
-                );
-
-                self.tokens.insert(
-                    to_ident_key(&call_path_binding.inner.suffix.suffix),
-                    token.clone(),
-                );
-
-                for type_arg in &call_path_binding.type_arguments {
-                    self.collect_type_arg(type_arg, &token);
-                }
-
-                for exp in args {
-                    self.handle_expression(exp);
-                }
+                path_expr.parse(ctx);
             }
             ExpressionKind::DelineatedPath(delineated_path_expression) => {
-                let DelineatedPathExpression {
-                    call_path_binding,
-                    args,
-                } = &**delineated_path_expression;
-                for ident in &call_path_binding.inner.prefixes {
-                    self.tokens.insert(
-                        to_ident_key(ident),
-                        Token::from_parsed(
-                            AstToken::Expression(expression.clone()),
-                            SymbolKind::Enum,
-                        ),
-                    );
-                }
-
-                let token = Token::from_parsed(
-                    AstToken::Expression(expression.clone()),
-                    SymbolKind::Variant,
-                );
-
-                self.tokens
-                    .insert(to_ident_key(&call_path_binding.inner.suffix), token.clone());
-
-                for type_arg in &call_path_binding.type_arguments {
-                    self.collect_type_arg(type_arg, &token);
-                }
-
-                for exp in args {
-                    self.handle_expression(exp);
-                }
+                delineated_path_expression.parse(ctx);
             }
             ExpressionKind::AbiCast(abi_cast_expression) => {
-                let AbiCastExpression { abi_name, address } = &**abi_cast_expression;
-                for ident in &abi_name.prefixes {
-                    self.tokens.insert(
-                        to_ident_key(ident),
-                        Token::from_parsed(
-                            AstToken::Expression(expression.clone()),
-                            SymbolKind::Module,
-                        ),
-                    );
-                }
-                self.tokens.insert(
-                    to_ident_key(&abi_name.suffix),
-                    Token::from_parsed(AstToken::Expression(expression.clone()), SymbolKind::Trait),
-                );
-                self.handle_expression(address);
+                abi_cast_expression.parse(ctx);
             }
             ExpressionKind::ArrayIndex(ArrayIndexExpression { prefix, index, .. }) => {
-                self.handle_expression(prefix);
-                self.handle_expression(index);
+                prefix.parse(ctx);
+                index.parse(ctx);
             }
-            ExpressionKind::StorageAccess(StorageAccessExpression { field_names, .. }) => {
-                for field in field_names {
-                    self.tokens.insert(
-                        to_ident_key(field),
-                        Token::from_parsed(
-                            AstToken::Expression(expression.clone()),
+            ExpressionKind::StorageAccess(storage_access_exp) => {
+                storage_access_exp
+                    .field_names
+                    .iter()
+                    .for_each(|field_name| {
+                        ctx.tokens.insert_parsed(
+                            field_name.clone(),
+                            AstToken::Ident(field_name.clone()),
                             SymbolKind::Field,
-                        ),
-                    );
-                }
+                        );
+                    });
             }
             ExpressionKind::IntrinsicFunction(IntrinsicFunctionExpression {
                 name,
                 kind_binding,
                 arguments,
             }) => {
-                self.tokens.insert(
-                    to_ident_key(name),
-                    Token::from_parsed(
-                        AstToken::Intrinsic(kind_binding.inner.clone()),
-                        SymbolKind::Function,
-                    ),
+                ctx.tokens.insert_parsed(
+                    name.clone(),
+                    AstToken::Intrinsic(kind_binding.inner.clone()),
+                    SymbolKind::Function,
                 );
-
-                for argument in arguments {
-                    self.handle_expression(argument);
-                }
+                arguments.iter().for_each(|argument| argument.parse(ctx));
             }
-            ExpressionKind::WhileLoop(WhileLoopExpression {
-                body, condition, ..
-            }) => self.handle_while_loop(body, condition),
-            // TODO: collect these tokens as keywords once the compiler returns the span
+            ExpressionKind::WhileLoop(while_exp) => {
+                while_exp.body.parse(ctx);
+                while_exp.condition.parse(ctx);
+            }
+            // These keyword tokens are already collected in during lexing traversal.
             ExpressionKind::Break | ExpressionKind::Continue => {}
             ExpressionKind::Reassignment(reassignment) => {
-                self.handle_expression(&reassignment.rhs);
-
-                match &reassignment.lhs {
-                    ReassignmentTarget::VariableExpression(exp) => {
-                        self.handle_expression(exp);
-                    }
-                    ReassignmentTarget::StorageField(idents) => {
-                        for ident in idents {
-                            self.tokens.insert(
-                                to_ident_key(ident),
-                                Token::from_parsed(
-                                    AstToken::Reassignment(reassignment.clone()),
-                                    SymbolKind::Field,
-                                ),
-                            );
-                        }
-                    }
-                }
+                reassignment.parse(ctx);
             }
-            ExpressionKind::Return(expr) => self.handle_expression(expr),
+            ExpressionKind::Return(expr) => expr.parse(ctx),
         }
     }
+}
 
-    fn handle_while_loop(&self, body: &CodeBlock, condition: &Expression) {
-        self.handle_expression(condition);
-        for node in &body.contents {
-            self.traverse_node(node);
-        }
-    }
+impl Parse for VariableDeclaration {
+    fn parse(&self, ctx: &ParseContext) {
+        // Don't collect tokens if the ident's name contains __tuple_ || __match_return_var_name_
+        // The individual elements are handled in the subsequent VariableDeclaration's
+        if !self.name.as_str().contains(TUPLE_NAME_PREFIX)
+            && !self.name.as_str().contains(MATCH_RETURN_VAR_NAME_PREFIX)
+        {
+            let symbol_kind = symbol_kind_from_ident(&self.name);
 
-    fn collect_type_arg(&self, type_argument: &TypeArgument, token: &Token) {
-        let mut token = token.clone();
-        let type_info = self.type_engine.get(type_argument.type_id);
-        match &type_info {
-            TypeInfo::Array(type_arg, length) => {
-                token.kind = SymbolKind::NumericLiteral;
-                self.tokens
-                    .insert(to_ident_key(&Ident::new(length.span())), token.clone());
-                self.collect_type_arg(type_arg, &token);
-            }
-            TypeInfo::Tuple(type_arguments) => {
-                for type_arg in type_arguments {
-                    self.collect_type_arg(type_arg, &token);
-                }
-            }
-            _ => {
-                let symbol_kind = type_info_to_symbol_kind(self.type_engine, &type_info);
-                token.kind = symbol_kind;
-                token.type_def = Some(TypeDefinition::TypeId(type_argument.type_id));
-                self.tokens
-                    .insert(to_ident_key(&Ident::new(type_argument.span.clone())), token);
+            // We want to use the span from variable.name to construct a
+            // new Ident as the name_override_opt can be set to one of the
+            // const prefixes and not the actual token name.
+            ctx.tokens.insert_parsed(
+                Ident::new(self.name.span()),
+                AstToken::VariableDeclaration(self.clone()),
+                symbol_kind,
+            );
+
+            if self.type_ascription_span.is_some() {
+                self.type_ascription.parse(ctx);
             }
         }
+        self.body.parse(ctx);
     }
+}
 
-    fn collect_scrutinee(&self, scrutinee: &Scrutinee) {
-        match scrutinee {
+impl Parse for FunctionDeclaration {
+    fn parse(&self, ctx: &ParseContext) {
+        ctx.tokens.insert_parsed(
+            self.name.clone(),
+            AstToken::FunctionDeclaration(self.clone()),
+            SymbolKind::Function,
+        );
+        self.body.contents.iter().for_each(|node| node.parse(ctx));
+        self.parameters
+            .iter()
+            .for_each(|func_param| func_param.parse(ctx));
+        self.type_parameters
+            .iter()
+            .for_each(|type_param| type_param.parse(ctx));
+        self.return_type.parse(ctx);
+        self.attributes.parse(ctx);
+    }
+}
+
+impl Parse for TraitDeclaration {
+    fn parse(&self, ctx: &ParseContext) {
+        ctx.tokens.insert_parsed(
+            self.name.clone(),
+            AstToken::TraitDeclaration(self.clone()),
+            SymbolKind::Trait,
+        );
+        self.interface_surface
+            .iter()
+            .for_each(|trait_fn| trait_fn.parse(ctx));
+        self.methods
+            .iter()
+            .for_each(|func_decl| func_decl.parse(ctx));
+        self.supertraits
+            .iter()
+            .for_each(|supertrait| supertrait.parse(ctx));
+    }
+}
+
+impl Parse for StructDeclaration {
+    fn parse(&self, ctx: &ParseContext) {
+        ctx.tokens.insert_parsed(
+            self.name.clone(),
+            AstToken::StructDeclaration(self.clone()),
+            SymbolKind::Struct,
+        );
+        self.fields.iter().for_each(|field| field.parse(ctx));
+        self.type_parameters.iter().for_each(|type_param| {
+            type_param.parse(ctx);
+        });
+        self.attributes.parse(ctx);
+    }
+}
+
+impl Parse for EnumDeclaration {
+    fn parse(&self, ctx: &ParseContext) {
+        ctx.tokens.insert_parsed(
+            self.name.clone(),
+            AstToken::EnumDeclaration(self.clone()),
+            SymbolKind::Enum,
+        );
+        self.type_parameters.iter().for_each(|type_param| {
+            type_param.parse(ctx);
+        });
+        self.variants.iter().for_each(|variant| variant.parse(ctx));
+        self.attributes.parse(ctx);
+    }
+}
+
+impl Parse for ImplTrait {
+    fn parse(&self, ctx: &ParseContext) {
+        self.trait_name.prefixes.iter().for_each(|prefix| {
+            ctx.tokens.insert_parsed(
+                prefix.clone(),
+                AstToken::Ident(prefix.clone()),
+                SymbolKind::Module,
+            );
+        });
+        ctx.tokens.insert_parsed(
+            self.trait_name.suffix.clone(),
+            AstToken::ImplTrait(self.clone()),
+            SymbolKind::Trait,
+        );
+
+        self.type_implementing_for.parse(ctx);
+        self.impl_type_parameters.iter().for_each(|type_param| {
+            type_param.parse(ctx);
+        });
+        self.functions
+            .iter()
+            .for_each(|func_decl| func_decl.parse(ctx));
+    }
+}
+
+impl Parse for ImplSelf {
+    fn parse(&self, ctx: &ParseContext) {
+        self.type_implementing_for.parse(ctx);
+        self.impl_type_parameters.iter().for_each(|type_param| {
+            type_param.parse(ctx);
+        });
+        self.functions
+            .iter()
+            .for_each(|func_decl| func_decl.parse(ctx));
+    }
+}
+
+impl Parse for AbiDeclaration {
+    fn parse(&self, ctx: &ParseContext) {
+        ctx.tokens.insert_parsed(
+            self.name.clone(),
+            AstToken::AbiDeclaration(self.clone()),
+            SymbolKind::Trait,
+        );
+        self.interface_surface
+            .iter()
+            .for_each(|trait_fn| trait_fn.parse(ctx));
+        self.attributes.parse(ctx);
+    }
+}
+
+impl Parse for ConstantDeclaration {
+    fn parse(&self, ctx: &ParseContext) {
+        ctx.tokens.insert_parsed(
+            self.name.clone(),
+            AstToken::ConstantDeclaration(self.clone()),
+            SymbolKind::Const,
+        );
+        self.type_ascription.parse(ctx);
+        self.value.parse(ctx);
+        self.attributes.parse(ctx);
+    }
+}
+
+impl Parse for StorageDeclaration {
+    fn parse(&self, ctx: &ParseContext) {
+        self.fields.iter().for_each(|field| field.parse(ctx));
+        self.attributes.parse(ctx);
+    }
+}
+
+impl Parse for FunctionApplicationExpression {
+    fn parse(&self, ctx: &ParseContext) {
+        // Don't collect applications of desugared operators due to mismatched ident lengths.
+        if !desugared_op(&self.call_path_binding.inner.prefixes) {
+            self.call_path_binding
+                .inner
+                .prefixes
+                .iter()
+                .for_each(|prefix| {
+                    ctx.tokens.insert_parsed(
+                        prefix.clone(),
+                        AstToken::Ident(prefix.clone()),
+                        SymbolKind::Module,
+                    );
+                });
+            ctx.tokens.insert_parsed(
+                self.call_path_binding.inner.suffix.clone(),
+                AstToken::FunctionApplicationExpression(self.clone()),
+                SymbolKind::Function,
+            );
+            self.call_path_binding
+                .type_arguments
+                .iter()
+                .for_each(|type_arg| {
+                    type_arg.parse(ctx);
+                });
+        }
+        self.arguments.iter().for_each(|arg| arg.parse(ctx));
+    }
+}
+
+impl Parse for CodeBlock {
+    fn parse(&self, ctx: &ParseContext) {
+        self.contents.iter().for_each(|node| node.parse(ctx));
+    }
+}
+
+impl Parse for TypeInfo {
+    fn parse(&self, ctx: &ParseContext) {
+        todo!();
+    }
+}
+
+impl Parse for TypeArgument {
+    fn parse(&self, ctx: &ParseContext) {
+        let type_info = ctx.engines.te().get(self.type_id);
+        type_info.parse(ctx);
+    }
+}
+
+impl Parse for TypeParameter {
+    fn parse(&self, ctx: &ParseContext) {
+        ctx.tokens.insert_parsed(
+            self.name_ident.clone(),
+            AstToken::TypeParameter(self.clone()),
+            SymbolKind::TypeParameter,
+        );
+    }
+}
+
+impl Parse for StorageField {
+    fn parse(&self, ctx: &ParseContext) {
+        ctx.tokens.insert_parsed(
+            self.name.clone(),
+            AstToken::StorageField(self.clone()),
+            SymbolKind::Field,
+        );
+        self.type_info.parse(ctx);
+        self.initializer.parse(ctx);
+        self.attributes.parse(ctx);
+    }
+}
+
+impl Parse for FunctionParameter {
+    fn parse(&self, ctx: &ParseContext) {
+        ctx.tokens.insert_parsed(
+            self.name.clone(),
+            AstToken::FunctionParameter(self.clone()),
+            SymbolKind::ValueParam,
+        );
+        self.type_info.parse(ctx);
+    }
+}
+
+impl Parse for TraitFn {
+    fn parse(&self, ctx: &ParseContext) {
+        ctx.tokens.insert_parsed(
+            self.name.clone(),
+            AstToken::TraitFn(self.clone()),
+            SymbolKind::Function,
+        );
+        self.parameters.iter().for_each(|param| param.parse(ctx));
+        self.return_type.parse(ctx);
+        self.attributes.parse(ctx);
+    }
+}
+
+impl Parse for StructField {
+    fn parse(&self, ctx: &ParseContext) {
+        ctx.tokens.insert_parsed(
+            self.name.clone(),
+            AstToken::StructField(self.clone()),
+            SymbolKind::Field,
+        );
+        self.type_info.parse(ctx);
+        self.attributes.parse(ctx);
+    }
+}
+
+impl Parse for StructExpression {
+    fn parse(&self, ctx: &ParseContext) {
+        self.call_path_binding
+            .inner
+            .prefixes
+            .iter()
+            .for_each(|prefix| {
+                ctx.tokens.insert_parsed(
+                    prefix.clone(),
+                    AstToken::Ident(prefix.clone()),
+                    SymbolKind::Struct,
+                );
+            });
+        ctx.tokens.insert_parsed(
+            self.call_path_binding.inner.suffix.clone(),
+            AstToken::StructExpression(self.clone()),
+            SymbolKind::Struct,
+        );
+        self.call_path_binding
+            .type_arguments
+            .iter()
+            .for_each(|type_arg| {
+                type_arg.parse(ctx);
+            });
+        self.fields.iter().for_each(|field| field.parse(ctx));
+    }
+}
+
+impl Parse for StructExpressionField {
+    fn parse(&self, ctx: &ParseContext) {
+        ctx.tokens.insert_parsed(
+            self.name.clone(),
+            AstToken::StructExpressionField(self.clone()),
+            SymbolKind::Field,
+        );
+        self.value.parse(ctx);
+    }
+}
+
+impl Parse for ArrayExpression {
+    fn parse(&self, ctx: &ParseContext) {
+        self.contents.iter().for_each(|expr| expr.parse(ctx));
+        if let Some(length_span) = &self.length_span {
+            let ident = Ident::new(length_span.clone());
+            ctx.tokens.insert_parsed(
+                ident.clone(),
+                AstToken::Ident(ident.clone()),
+                SymbolKind::NumericLiteral,
+            );
+        }
+    }
+}
+
+impl Parse for EnumVariant {
+    fn parse(&self, ctx: &ParseContext) {
+        ctx.tokens.insert_parsed(
+            self.name.clone(),
+            AstToken::EnumVariant(self.clone()),
+            SymbolKind::Variant,
+        );
+        self.type_info.parse(ctx);
+        self.attributes.parse(ctx);
+    }
+}
+
+impl Parse for Scrutinee {
+    fn parse(&self, ctx: &ParseContext) {
+        match self {
             Scrutinee::CatchAll { .. } => (),
             Scrutinee::Literal { ref value, span } => {
-                let token = Token::from_parsed(
-                    AstToken::Scrutinee(scrutinee.clone()),
+                ctx.tokens.insert_parsed(
+                    Ident::new(span.clone()),
+                    AstToken::Scrutinee(self.clone()),
                     literal_to_symbol_kind(value),
                 );
-                self.tokens
-                    .insert(to_ident_key(&Ident::new(span.clone())), token);
             }
             Scrutinee::Variable { name, .. } => {
-                let token = Token::from_parsed(
-                    AstToken::Scrutinee(scrutinee.clone()),
+                ctx.tokens.insert_parsed(
+                    name.clone(),
+                    AstToken::Scrutinee(self.clone()),
                     SymbolKind::Variable,
                 );
-                self.tokens.insert(to_ident_key(name), token);
             }
             Scrutinee::StructScrutinee {
                 struct_name,
                 fields,
                 ..
             } => {
-                let token =
-                    Token::from_parsed(AstToken::Scrutinee(scrutinee.clone()), SymbolKind::Struct);
-                self.tokens.insert(to_ident_key(struct_name), token);
-
-                for field in fields {
-                    let token = Token::from_parsed(
-                        AstToken::Scrutinee(scrutinee.clone()),
-                        SymbolKind::Field,
-                    );
-                    if let StructScrutineeField::Field {
-                        field, scrutinee, ..
-                    } = field
-                    {
-                        self.tokens.insert(to_ident_key(field), token);
-
-                        if let Some(scrutinee) = scrutinee {
-                            self.collect_scrutinee(scrutinee);
-                        }
-                    }
-                }
+                ctx.tokens.insert_parsed(
+                    struct_name.clone(),
+                    AstToken::Scrutinee(self.clone()),
+                    SymbolKind::Struct,
+                );
+                fields.iter().for_each(|field| field.parse(ctx));
             }
             Scrutinee::EnumScrutinee {
                 call_path, value, ..
             } => {
-                for ident in &call_path.prefixes {
-                    let token = Token::from_parsed(
-                        AstToken::Scrutinee(scrutinee.clone()),
+                call_path.prefixes.iter().for_each(|prefix| {
+                    ctx.tokens.insert_parsed(
+                        prefix.clone(),
+                        AstToken::Ident(prefix.clone()),
                         SymbolKind::Enum,
                     );
-                    self.tokens.insert(to_ident_key(ident), token);
-                }
-
-                let token =
-                    Token::from_parsed(AstToken::Scrutinee(scrutinee.clone()), SymbolKind::Variant);
-                self.tokens.insert(to_ident_key(&call_path.suffix), token);
-
-                self.collect_scrutinee(value);
+                });
+                ctx.tokens.insert_parsed(
+                    call_path.suffix.clone(),
+                    AstToken::Scrutinee(self.clone()),
+                    SymbolKind::Variant,
+                );
+                value.parse(ctx);
             }
             Scrutinee::Tuple { elems, .. } => {
-                for elem in elems {
-                    self.collect_scrutinee(elem);
-                }
+                elems.iter().for_each(|elem| elem.parse(ctx));
             }
             Scrutinee::Error { .. } => {
                 // FIXME: Left for @JoshuaBatty to use.
             }
         }
     }
+}
 
-    fn collect_type_info_token(
-        &self,
-        token: &Token,
-        type_info: &TypeInfo,
-        type_span: Option<Span>,
-        symbol_kind: Option<SymbolKind>,
-    ) {
-        let mut token = token.clone();
-        match symbol_kind {
-            Some(kind) => token.kind = kind,
-            None => token.kind = type_info_to_symbol_kind(self.type_engine, type_info),
-        }
-
-        match type_info {
-            TypeInfo::Str(length) => {
-                self.tokens
-                    .insert(to_ident_key(&Ident::new(length.span())), token);
-            }
-            TypeInfo::Array(type_arg, length) => {
-                token.kind = SymbolKind::NumericLiteral;
-                self.tokens
-                    .insert(to_ident_key(&Ident::new(length.span())), token.clone());
-                self.collect_type_arg(type_arg, &token);
-            }
-            TypeInfo::Tuple(type_arguments) => {
-                for type_arg in type_arguments {
-                    self.collect_type_arg(type_arg, &token);
-                }
-            }
-            TypeInfo::Custom {
-                name,
-                type_arguments,
-            } => {
-                token.type_def = Some(TypeDefinition::Ident(name.clone()));
-                self.tokens.insert(to_ident_key(name), token.clone());
-                if let Some(type_arguments) = type_arguments {
-                    for type_arg in type_arguments {
-                        self.collect_type_arg(type_arg, &token);
-                    }
-                }
-            }
-            _ => {
-                if let Some(type_span) = type_span {
-                    self.tokens
-                        .insert(to_ident_key(&Ident::new(type_span)), token);
-                }
+impl Parse for StructScrutineeField {
+    fn parse(&self, ctx: &ParseContext) {
+        if let StructScrutineeField::Field {
+            field, scrutinee, ..
+        } = self
+        {
+            ctx.tokens.insert_parsed(
+                field.clone(),
+                AstToken::StructScrutineeField(self.clone()),
+                SymbolKind::Field,
+            );
+            if let Some(scrutinee) = scrutinee {
+                scrutinee.parse(ctx);
             }
         }
     }
+}
 
-    fn collect_function_parameter(&self, parameter: &FunctionParameter) {
-        let token = Token::from_parsed(
-            AstToken::FunctionParameter(parameter.clone()),
-            SymbolKind::ValueParam,
-        );
-        self.tokens
-            .insert(to_ident_key(&parameter.name), token.clone());
+impl Parse for MethodApplicationExpression {
+    fn parse(&self, ctx: &ParseContext) {
+        let prefixes = match &self.method_name_binding.inner {
+            MethodName::FromType {
+                call_path_binding, ..
+            } => call_path_binding.inner.prefixes.clone(),
+            MethodName::FromTrait { call_path, .. } => call_path.prefixes.clone(),
+            _ => vec![],
+        };
 
-        self.collect_type_info_token(
-            &token,
-            &parameter.type_info,
-            Some(parameter.type_span.clone()),
-            None,
-        );
-    }
-
-    fn collect_trait_fn(&self, trait_fn: &TraitFn) {
-        let token = Token::from_parsed(AstToken::TraitFn(trait_fn.clone()), SymbolKind::Function);
-        self.tokens
-            .insert(to_ident_key(&trait_fn.name), token.clone());
-
-        for parameter in &trait_fn.parameters {
-            self.collect_function_parameter(parameter);
+        if let MethodName::FromType {
+            call_path_binding, ..
+        } = &self.method_name_binding.inner
+        {
+            let (type_info, ..) = &call_path_binding.inner.suffix;
+            type_info.parse(ctx);
         }
 
-        self.collect_type_info_token(
-            &token,
-            &trait_fn.return_type,
-            Some(trait_fn.return_type_span.clone()),
-            None,
-        );
-
-        trait_fn.attributes.parse(self.tokens);
+        // Don't collect applications of desugared operators due to mismatched ident lengths.
+        if !desugared_op(&prefixes) {
+            ctx.tokens.insert_parsed(
+                self.method_name_binding.inner.easy_name(),
+                AstToken::MethodApplicationExpression(self.clone()),
+                SymbolKind::Struct,
+            );
+        }
+        self.arguments.iter().for_each(|arg| arg.parse(ctx));
+        self.contract_call_params
+            .iter()
+            .for_each(|param| param.parse(ctx));
     }
+}
 
-    fn collect_type_parameter(&self, type_param: &TypeParameter, token: AstToken) {
-        self.tokens.insert(
-            to_ident_key(&type_param.name_ident),
-            Token::from_parsed(token, SymbolKind::TypeParameter),
+impl Parse for MatchBranch {
+    fn parse(&self, ctx: &ParseContext) {
+        self.scrutinee.parse(ctx);
+        self.result.parse(ctx);
+    }
+}
+
+impl Parse for AmbiguousPathExpression {
+    fn parse(&self, ctx: &ParseContext) {
+        self.call_path_binding
+            .inner
+            .prefixes
+            .iter()
+            .chain(std::iter::once(
+                &self.call_path_binding.inner.suffix.before.inner,
+            ))
+            .for_each(|ident| {
+                ctx.tokens.insert_parsed(
+                    ident.clone(),
+                    AstToken::Ident(ident.clone()),
+                    SymbolKind::Enum,
+                );
+            });
+        ctx.tokens.insert_parsed(
+            self.call_path_binding.inner.suffix.suffix.clone(),
+            AstToken::AmbiguousPathExpression(self.clone()),
+            SymbolKind::Variant,
+        );
+        self.call_path_binding
+            .type_arguments
+            .iter()
+            .for_each(|type_arg| {
+                type_arg.parse(ctx);
+            });
+        self.args.iter().for_each(|arg| arg.parse(ctx));
+    }
+}
+
+impl Parse for DelineatedPathExpression {
+    fn parse(&self, ctx: &ParseContext) {
+        self.call_path_binding
+            .inner
+            .prefixes
+            .iter()
+            .for_each(|ident| {
+                ctx.tokens.insert_parsed(
+                    ident.clone(),
+                    AstToken::Ident(ident.clone()),
+                    SymbolKind::Enum,
+                );
+            });
+        ctx.tokens.insert_parsed(
+            self.call_path_binding.inner.suffix.clone(),
+            AstToken::DelineatedPathExpression(self.clone()),
+            SymbolKind::Variant,
+        );
+        self.call_path_binding
+            .type_arguments
+            .iter()
+            .for_each(|type_arg| {
+                type_arg.parse(ctx);
+            });
+        self.args.iter().for_each(|arg| arg.parse(ctx));
+    }
+}
+
+impl Parse for AbiCastExpression {
+    fn parse(&self, ctx: &ParseContext) {
+        self.abi_name.prefixes.iter().for_each(|ident| {
+            ctx.tokens.insert_parsed(
+                ident.clone(),
+                AstToken::Ident(ident.clone()),
+                SymbolKind::Module,
+            );
+        });
+        ctx.tokens.insert_parsed(
+            self.abi_name.suffix.clone(),
+            AstToken::AbiCastExpression(self.clone()),
+            SymbolKind::Trait,
+        );
+        self.address.parse(ctx);
+    }
+}
+
+impl Parse for ReassignmentExpression {
+    fn parse(&self, ctx: &ParseContext) {
+        self.rhs.parse(ctx);
+        match &self.lhs {
+            ReassignmentTarget::VariableExpression(exp) => {
+                exp.parse(ctx);
+            }
+            ReassignmentTarget::StorageField(idents) => {
+                idents.iter().for_each(|ident| {
+                    ctx.tokens.insert_parsed(
+                        ident.clone(),
+                        AstToken::ReassignmentExpression(self.clone()),
+                        SymbolKind::Field,
+                    );
+                });
+            }
+        }
+    }
+}
+
+impl Parse for Supertrait {
+    fn parse(&self, ctx: &ParseContext) {
+        ctx.tokens.insert_parsed(
+            self.name.suffix.clone(),
+            AstToken::Supertrait(self.clone()),
+            SymbolKind::Trait,
         );
     }
 }
 
 impl Parse for AttributesMap {
-    fn parse(&self, tokens: &TokenMap) {
+    fn parse(&self, ctx: &ParseContext) {
         self.iter()
             .filter(|(kind, ..)| **kind != AttributeKind::DocComment)
             .flat_map(|(.., attrs)| attrs)
             .for_each(|attribute| {
-                tokens.insert(
-                    to_ident_key(&attribute.name),
-                    Token::from_parsed(
-                        AstToken::Attribute(attribute.clone()),
-                        SymbolKind::DeriveHelper,
-                    ),
+                ctx.tokens.insert_parsed(
+                    attribute.name.clone(),
+                    AstToken::Attribute(attribute.clone()),
+                    SymbolKind::DeriveHelper,
                 );
             });
     }
@@ -951,5 +781,13 @@ fn literal_to_symbol_kind(value: &Literal) -> SymbolKind {
         Literal::String(..) => SymbolKind::StringLiteral,
         Literal::B256(..) => SymbolKind::ByteLiteral,
         Literal::Boolean(..) => SymbolKind::BoolLiteral,
+    }
+}
+
+fn symbol_kind_from_ident(ident: &Ident) -> SymbolKind {
+    if ident.as_str().contains(DESTRUCTURE_PREFIX) {
+        SymbolKind::Struct
+    } else {
+        SymbolKind::Variable
     }
 }
